requirements.txt
discord-components==2.1.2
pytz==2021.3
mysql-connector-python==8.0.27
Flask==2.0.2
discord.py==1.7.3
discord~=1.7.3